+++ src/module/car/service/carService.js
@@ -1,3 +1,7 @@
+const Car = require('../entity/car');
+const CarIdNotDefinedError = require('../error/CarIdNotDefinedError');
+const CarNotDefinedError = require('../error/CarNotDefinedError');
+
 class CarService {
   constructor(carRepository) {
     this.repository = carRepository;
@@ -8,14 +12,26 @@ class CarService {
   }
 
   async getById(id) {
+    if (id === undefined) {
+      throw new CarIdNotDefinedError();
+    }
+
     return this.repository.getById(id);
   }
 
   async save(car) {
+    if (!(car instanceof Car)) {
+      throw new CarNotDefinedError();
+    }
+
     return this.repository.save(car);
   }
 
   async delete(id) {
+    if (id === undefined) {
+      throw new CarIdNotDefinedError();
+    }
+
     return this.repository.delete(id);
   }
 }
